% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum.R
\docType{methods}
\name{addData}
\alias{addData}
\alias{addData,simplecpt-method}
\title{Add more data from a vector}
\usage{
addData(obj, data)

\S4method{addData}{simplecpt}(obj, data)
}
\arguments{
\item{obj}{An object of the class "simplecpt".}

\item{data}{A numeric input vector.}
}
\value{
An object of the class "simplecpt".
}
\description{
Add more data from a vector
}
\examples{
require(simplecpt)
obj <- simplecpt()
obj <- addData(obj, rnorm(100,mean=1))

}
\references{
https://en.wikipedia.org/wiki/CUSUM
}
