% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum.R
\docType{methods}
\name{checkCpt}
\alias{checkCpt}
\alias{checkCpt,simplecpt-method}
\title{Check if any cusum statistic exceeds the threshold}
\usage{
checkCpt(obj, threshold)

\S4method{checkCpt}{simplecpt}(obj, threshold)
}
\arguments{
\item{obj}{An object of the class "simplecpt".}

\item{threshold}{A numeric value indicating the threshold for the cusum statistic.}
}
\value{
True if any of the cusum statistic exceeds the threshold.
}
\description{
Check if any cusum statistic exceeds the threshold
}
\examples{
require(simplecpt)
obj <- simplecpt()
obj <- addData(obj, c(1, 8, 3))
checkCpt(obj, 0.01)
checkCpt(obj, 100)

}
\references{
https://en.wikipedia.org/wiki/CUSUM
}
