% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum.R
\docType{class}
\name{simplecpt-class}
\alias{simplecpt-class}
\title{S4 class to detect changepoints}
\description{
Implementation of the CUSUM technique for seqential analysis in R
}
\section{Slots}{

\describe{
\item{\code{a}}{This is the variable that contains the numeric vector}
}}

\examples{
require(simplecpt)
obj <- simplecpt()
obj <- addData(obj, c(1, 8, 3))
checkCpt(obj, 0.01)
checkCpt(obj, 100)

}
\references{
Page, E.S. (1954). "Continuous Inspection Scheme". Biometrika. 41 (1/2): 100-115
}
